This is release version 1.5.7.
This release features a working version of the circuit simulator for fairly complex multiconverter circuits.

The focus on this release was in getting to a working circuit simulator after seeing it break over and over again
for circuits of increasing complexity. However, this version in most likelihood performs repeated calculations
that can be avoided and uses data structures that are too large and inefficient. The most important being the
LoopMap. The LoopMap will indicate the presence or absence of a branch in a loop. However, for large circuits,
a loop that will contain only a few branches will end up with an entry for every branch which is a sheer waste
of memory.