Made a very minor modification.
Decided to abandon this concept. Building a GUI interface out of Django seemed
like a neat idea but don't think it will work out. The configuration of a database
and the running of a server, even if it can be simplified for new Python users,
would still result in too many dependencies. If a GUI application needs to be created
it would have to be done using Tkinter, PyQt or something similar. It is reasonable
to expect a user to install Python and MatPlotlib but everything else should be minimal.
A database doesn't seem like the right way to go.