Continuing with the Django based user interface.
Reconsidered my previous decision to abandon the project. There may be ways to automate the server and database setup. So therefore, it may not be too hard for someone to configure. If Django provides a starting point for a decent user interface, further extensions can be planned later. Now the next step is to plan how the two projects will diverge - using command line and using Django. And also, how the functions will change that read the circuit schematics.