Submitting a basic simulation parameters form is done except for selecting the circuit schematics or the control code.
At this point a bifurcation needs to be performed between the two git repositories - the simulator and the cases. Up to now all the files of the simulator had to be in the same folder with the circuit files but now simulation files will be a database which will exist in only one folder. Therefore, a separate git folder will made on the laptop that will sync and push to the branch sim_with_django that will contain the web app files.