Continued with the simulation case form with a file field to upload a circuit file.
Spent a lot of time trying to extract the path of the circuit file on the local filesystem so that it can be set as the working directory. That just was not possible because that is how Django is designed. The idea is that a web app should not know the filesystem on a local computer for security reasons. The user should be able to upload files that end up in the server MEDIA_ROOT directory. This won't work for the simulator. The closest I could come so far was to let the user enter the working directory. The circuit file uploads are not really uploads but simply extracting the names of the circuit files which can then be appended to the working directory path to access the files later. Seems a little inelegant but for now will work.