Changed the relationship between SimulationCase and CircuitSchematics in terms of ForeignKey relationships. Also, edited the HTML form to return model ID.
I always get confused between how models should refer to each other when they are related through ForeignKey. When defining a model as a ForeignKey of another, there can only be one such definition. So, if there can be many CircuitSchematics for a single SimulationCase, SimulationCase must be defined as a ForeignKey for CircuitSchematics. Then, many CircuitSchematics models can be instantiated and pointed to a single SimulationCase model. Not the other way around.