Added the circuit files successfully and read them. However, the delete circuit is not working.
Still not sure about the being able to generate the full path names to the circuit files. So far working but need to ensure that it works cross-platform. When separating the folders in the path and trying to join them using the os.path.join function, it doesn't work because windows needs C:\ and not just c:. I wonder if this is platform specific.