Form validation on simulation parameters has been extended and checking for circuit spreadsheets being readable is partly done.
Checking has been added to ensure that simulation time limit is a positive number, data storage interval is greater than or equal to integration time step and if slicing output file option is chosen, the number of slices has to be greater than two. The check whether a circuit spreadsheet exists in the working directory and also whether it can be read is partly done. The HTML form doesn't render as conveniently as the simulation parameter form. Need to figure out the formatting of the error messages.