Fixed the problem of not being able to display error properly when circuit spreadsheet is not in the right directory.
Care needs to be taken when rendering a form and how many elements are there. If custom form rendering is needed, it is essential that we loop through item in a form and access the error message for each field. The next step is to read the circuits and instantiate objects for each circuit component found.