Functions have been cleaned up to look cleaner since the plan is to have the entire simulation in once page.
The next step is to read the circuit files and display any errors the same as with command line. The only question is about the exit function used in the command line version that exits the simulator when an error is detected. Will the exit cause the server to stop. Maybe not, but still can't take the chance. Therefore, every function will have to return an error list which will be displayed. This can then be ported to the web app.