The form has been split into a number of smaller forms to make data entry easier. More error checking has been implemented in the forms.
The simulation has another variable called sim_stage which contains a pointer about at what stage of data entry the simulation is. The sim_stage so far has 1 for simulation parameters and 2 for circuit spreadsheets but there will be many more stages. There will be a main page where the sim_stage produces an instruction for the user on what to do next for the simulation to continue. The next stage will be to start reading the circuit and processing it.