The form is almost complete with all error messages. The processing of the circuit has started.
Errors at every stage will be pointed out by the simulator. This is even with circuit schematics that have already been uploaded. Which has been acheived through model_to_dict since a form instantiated from a model object is an unbound form and error checking cannot be run on it.