Circuit errors in components and also in connectivity have been included in the view which passes back an error list.
The problem now is in how to handle the objects. Before, everything was in the circuit_solver file. But now it is in separate views and passed back and forth between views and templates. Therefore, either code that identifies components will need to be repeated or the components will need to be stored in a database and retrieved at any point of time. The database option seems cleaner and in the long run may be more efficient though for smaller circuits a database query will lead to more overhead. The instantiating of component objects will need to be done at the last stage when the simulation is being run.