The components read from the circuit schematics and entered into a database for future use.
Before, there was components_found which contained a list of all the components in a circuit. However, in a web app, this will not be enough as the dictionaries cannot be passed back and forth between views and templates. So, if data needs to be preserved it needs to be entered in a database. The template file also lists the errors in the circuit spreadsheet and specifies when it is ready to go further. Now, the next stage is to create models for every type of components and add them to either circuit schematic models or the simulation. Circuit schematic spreadsheets makes sense as the purpose behind these component models is to create forms that can be edited through circuit schematics.