Started with creating individual component models so as to create parameter forms.
The main challenge has been to create a generic interface to connect to each component model. The interface used has been component_objects and the instances of every type of component. A separate method has been created to create new component models, edit old component models and also return a list of existing component models to the webpage which will display forms. The next step will be to create the forms, the polarity information and include error checking with the forms.