Most of the forms are done with error messages and this includes checking polarity of components.
The major challenge was the design of components and objects. The next stage is in running the simulation and checking the results. After that plots using MatPlotLib can be used to view waveforms. About running the simulation, the best way of doing so I can think of is to add a field to the main database whether the simulation is running and how much is done. This will help to always keep track of the simulation. However, if the server stops or the script for some reason stops would the update still happen? And will this cause problems as the run bit will not be cleared.