Multiprocessing has been chosen for running the simulator loop in the background rather than multi threading.
Multithreading is tougher to control and does not seem to bring significant improvement due to the Global Interpretor Lock implemented in Python. Threading does not have a convenient kill or terminate function due to which the thread doesn't seem to die. Multiprcoessing however is very convenient to start and stop. By adding a global variable dictionary that has all the simulations that are run simultaneously, the process object can be terminated when needed to stop the simulation. A pause seems tougher to implement and is deferred for now.