Added the list of simulations page and started with the plot description using a separate model for circuit plots.
The plots are to be structured with respect to simulation cases. A simulation can have multiple circuit plots and therefore is a foregin key. Each simulation at the same time will have models for each element that can be plotted. These elements are meters (for now) and are related to the simulation case as a foreign key. At the same time, the plot elements can be in many circuit plots and in turn a circuit plot can have many plot items. So therefore, these are related as many to many relationships. However, they cannot be directly related because a plot element can have different scalings in different circuit plots. So there is an intermediary model called waveforms. A waveform has a circuit plot as a foregin key and is related to the plot elements as a many to many relationship.