The basic interface is working with control and also saving plots to png files.
The control has been interfaced with the simulator engine and now variable resistor is working. Will test the remaining controllable elements. Also, due to repeated crashes of the server due to using MatPlotLib to display the plot in a window a better method seems to save it as a png file. There doesn't seem to be any stability issues this way.