Releasing version 1.5.8.
There was a bug in the previous version. Before solving the system, the loops are rearranged so that the loop is always through the path of minimal resistance. So, for every nonstiff loop, it is checked whether a branch has parallel branches. If so, the branch with the least resistance is chosen for the loop. However, there has to be a check whether this loop exists and if so not to replace the branch as this will result in a reptition of loops. This replacement was being performed because of which the loops cancelled out and the system was underrepresented.