Releasing version 2.0.3 and 1.5.8.
Bringing the changes in CLI 1.5.8 into the Webapp. While solving loop
equations, the loops are adjusted so that a loop passes through the
branches of least resistance. So in the case of parallel branches, the
branch with least resistance is chosen. But a check needs to be made
whether such a loop already exists and if so not to perform the
replacement as this will lead to duplicate loops.