Version 1.5.9 released.
There was a bug in the loop solver. There is a function
approximate_nonstiff_loops that takes the set of nonstiff loops that
have been generated by remove_stiffness. All the loops with more than 2
branches are edited so that if a branch has parallel branches, the
branch with the minimum resistance is the one in the loop. This ensures
that a loop has the path of minimum resistance. After this all loops
with two branches are deleted and new branches with two loops are added.
The problem is that this results in redundant loops. So, all these
functions are brought into remove_stiffness. First, loops between
parallel branches are added as these represent the system the best. Then
with every loop manipulation, it is checked if the branches have minimum
possible resistance. In the end, extra loops are deleted. So, it needs
to be seen if the manipulation of the non stiff loops will always result
in independent loop and if there is any potential of dependent loops how
to check for them.